created CreditManager script
