+++ Assets/Scripts/CreditManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f150e54069f81c4f828e99b4f6a8343
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
