+++ Assets/Scripts/LevelManager.cs
@@ -7,14 +7,11 @@ public class LevelManager : MonoBehaviour
     public static LevelManager LM;
     public Transform[] TurningPoints;
     public GameObject enermy;
-    public int credits;
 
     void Start()
     {
         LM = this;
         InvokeRepeating("spawnMouse", 0, 5);
-        //some arbitrary number for now.
-        credits = 50;
     }
 
     private void spawnMouse()
@@ -22,23 +19,5 @@ public class LevelManager : MonoBehaviour
         Instantiate(enermy, TurningPoints[0].position, transform.rotation);
     }
 
-    public void increaseMoney(int amount)
-    {
-        credits += amount;
-    }
-
-    public bool spendCredits(int amount)
-    {
-        if (credits >= amount)
-        {
-            credits -= amount;
-            return true;
-        }
-        else
-        {
-            //console log for now.
-            Debug.Log("no money");
-            return false;
-        }
-    }
+    
 }
